refactor: moved item request validation from server to gateway, fix bug
